// frontend-react/src/api/models/adminTrackDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ArtistSummaryDto } from './artistSummaryDto';
import type { AlbumSummaryDto } from './albumSummaryDto';
import type { PlaybackLinkDto } from './playbackLinkDto';

export interface AdminTrackDto {
  id?: string;
  title?: string;
  isrc?: string;
  durationMs?: number;
  tempoBpm?: number;
  createdAt?: string;
  processingStatus?: string;
  isFlagged?: boolean;
  flagReason?: string;
  hasVocals?: boolean;
  musicGenre?: string;
  genreConfidence?: number;
  artists?: ArtistSummaryDto[];
  album?: AlbumSummaryDto;
  playbackLinks?: PlaybackLinkDto[];
  danceStyle?: string;
  subStyle?: string;
  tempoCategory?: string;
  confidence?: number;
}